optimize code. Default output empty id -1

+++ mybatis-model-helper/src/main/java/com/javaoffers/batis/modelhelper/convert/Serializable2IdConvert.java
@@ -30,7 +30,7 @@ class IdImpl implements Id {
     Serializable srcValue;
 
     @Override
-    public int toByte() {
+    public byte toByte() {
         return convert.converterObject(byte.class, srcValue);
     }
